cformat.service: sandbox the service
Many options that I’d like to have are not available in user services
since the user service manager itself lacks privileges to set them up;
when cformat can run as root (when ceylon.formatter/1.3.2, with
ceylon/ceylon.formatter#132, is released), I’ll add those in an extra
config file that can be used for a system service.

I’m still pretty pleased with the result. Several powerful sandboxing
options are still available in the user manager, and if there is any
code execution vulnerability in the formatter and/or JVM, this should
limit the impact quite a bit. Incomplete list of things the formatter
cannot do:

- Talk to C&C servers. It has no socket(2) syscall and also no address
  families other than AF_UNIX. (That last part is probably redundant
  given the first, but whatever.)
- Remove files or directories (though it can still truncate files),
  create symbolic or hard links, rename files, change permissions or
  attributes. The system call filter includes only a small subset of the
  @file-system set.
- Execute any external commands. No fork(2), and no more than 1 task.
  (Unfortunately, we can’t block execve(2), since systemd needs it to
  execute the daemon process, but the daemon can only replace itself
  with another process, which is still subject to the same limitations.
  Also, the new process would not gain privileges even if it has the
  setuid bit set, thanks to NoNewPrivileges=yes.)